feat: migrated Root stake claim settings to Portfolio asset toolbar
